feat: commonJS refactored into ECMAScript modules
